requirements.txt
httpx
alive_progress
python-dotenv